% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{standards}
\alias{standards}
\title{Available Structured Query Language languages or standards available to use
in SQL formatting.}
\format{
A vector of four character strings.
}
\usage{
standards
}
\description{
Available Structured Query Language languages or standards available to use
in SQL formatting.
}
\keyword{datasets}
